+++ vllm_ascend/distributed/mooncake_connector.py
@@ -4,6 +4,7 @@ import hashlib
 import math
 import queue
 import random
+import socket
 import struct
 import threading
 import time
@@ -17,8 +18,9 @@ import msgspec
 import numpy as np
 import numpy.typing as npt
 import torch
+import torch_npu
 import zmq
-from mooncake.engine import TransferEngine  # type: ignore
+from mooncake.engine import TransferEngine
 from vllm.config import VllmConfig
 from vllm.distributed.kv_transfer.kv_connector.v1.base import (
     KVConnectorBase_V1, KVConnectorMetadata, KVConnectorRole)
@@ -38,6 +40,7 @@ if TYPE_CHECKING:
 
 GET_META_MSG = b"get_meta_msg"
 DONE_RECVING_MSG = b"done_recving_msg"
+KV_CACHE_DTYPE = torch.bfloat16
 
 
 class MooncakeAgentMetadata(msgspec.Struct, omit_defaults=True, dict=True):
@@ -119,7 +122,8 @@ class KVCacheSendingThread(threading.Thread):
     def __init__(self, tp_rank: int, decode_tp_size: int, local_engine_id: str,
                  side_channel_host: str, side_channel_port: int,
                  metadata: MooncakeAgentMetadata,
-                 ready_event: threading.Event):
+                 ready_event: threading.Event,
+                 kv_caches: dict[str, Any]):
         super().__init__(daemon=True, name="KVCacheSendingThread")
         self.tp_rank = tp_rank
         self.decode_tp_size = decode_tp_size
@@ -128,6 +132,7 @@ class KVCacheSendingThread(threading.Thread):
         self.side_channel_port = side_channel_port
         self.metadata = metadata
         self.ready_event = ready_event
+        self.kv_caches = kv_caches
 
         self.task_tracker = KVCacheTaskTracker()
 
@@ -159,7 +164,7 @@ class KVCacheSendingThread(threading.Thread):
         handshake_port = self.side_channel_port + self.tp_rank
         path = make_zmq_path("tcp", self.side_channel_host, handshake_port)
         logger.info("Starting listening on path: %s", path)
-        with zmq_ctx(zmq.ROUTER, path) as sock:  # type: ignore
+        with zmq_ctx(zmq.ROUTER, path) as sock:
             self.ready_event.set()
             decoder = msgspec.msgpack.Decoder(type=tuple)
             while True:
@@ -187,11 +192,10 @@ class KVCacheSendingThread(threading.Thread):
                         while True:
                             try:
                                 # Send ACK to the sender.
-                                sock.send_multipart(
-                                    (identity, b"", b"ACK"),
-                                    flags=zmq.NOBLOCK)  # type: ignore
+                                sock.send_multipart((identity, b"", b"ACK"),
+                                                    flags=zmq.NOBLOCK)
                                 break
-                            except zmq.Again:  # type: ignore
+                            except zmq.Again:
                                 # If the socket is not ready, retry sending.
                                 logger.debug(
                                     "Socket not ready, retrying to send ACK for "
@@ -211,7 +215,7 @@ class KVCacheRecvingThread(threading.Thread):
     def __init__(self, tp_rank: int, tp_size: int, engine: TransferEngine,
                  local_engine_id: str, local_handshake_port: int,
                  local_kv_caches_base_addr: list[int], block_len: list[int],
-                 ready_event: threading.Event):
+                 ready_event: threading.Event, vllm_config: VllmConfig, kv_caches: dict[str, Any]):
         super().__init__(daemon=True, name="KVCacheRecvingThread")
         self.tp_rank = tp_rank
         self.tp_size = tp_size
@@ -228,27 +232,30 @@ class KVCacheRecvingThread(threading.Thread):
         self.remote_te_port: dict[str, dict[int, int]] = \
             defaultdict(dict)
         self.block_len = block_len
-        # TODO(jianzs): find a better way to detect MLA.
-        self.use_mla = len(block_len) == 2
 
-        self.request_queue: queue.Queue[Any] = queue.Queue()
+        self.request_queue = queue.Queue()
         # TODO(jianzs): make this configurable
-        self.executor = ThreadPoolExecutor(max_workers=32)
+        max_workers = getattr(envs_ascend, 'MAX_TRANSFER_WORKERS', 32)
+        self.executor = ThreadPoolExecutor(max_workers=max_workers)
 
         self.task_tracker = KVCacheTaskTracker()
 
         self.encoder = msgspec.msgpack.Encoder()
         self.decoder = msgspec.msgpack.Decoder(MooncakeAgentMetadata)
         self.remote_sockets_lock = threading.Lock()
-        self.remote_sockets: dict[  # type: ignore
-            str, deque[zmq.Socket]] = defaultdict(  # type: ignore
-                deque)
-        self.remote_poller = zmq.Poller()  # type: ignore
+        self.remote_sockets: dict[str, deque[zmq.Socket]] = defaultdict(deque)
+        self.remote_poller = zmq.Poller()
         self.timeout = 1.0  # seconds
+
+        self.vllm_config = vllm_config
+        self.model_config = self.vllm_config.model_config
+        self.num_key_value_heads = self.model_config.hf_config.num_key_value_heads
+        self.kv_caches = kv_caches
 
     def add_request(self, request_id: str, local_block_ids: list[int],
                     remote_block_ids: list[int], remote_engine_id: str,
-                    remote_host: str, remote_handshake_port: int):
+                    remote_host: str, remote_handshake_port: int,
+                    offset: int, num_need_pulls: int):
         """Add a new request to the queue for processing."""
         logger.debug(f"Adding request {request_id} to the queue.")
         self.request_queue.put({
@@ -258,7 +265,9 @@ class KVCacheRecvingThread(threading.Thread):
             "remote_engine_id": remote_engine_id,
             "remote_host": remote_host,
             "remote_handshake_port": remote_handshake_port,
-        })
+            "offset": offset,
+            "num_need_pulls": num_need_pulls
+            })
 
     def get_and_clear_finished_requests(self) -> set[str]:
         """
@@ -286,6 +295,8 @@ class KVCacheRecvingThread(threading.Thread):
         request_id = req_meta["request_id"]
         remote_host = req_meta["remote_host"]
         remote_handshake_port = req_meta["remote_handshake_port"]
+        offset = req_meta["offset"]
+        num_need_pulls = req_meta["num_need_pulls"]
 
         try:
             logger.debug(
@@ -297,12 +308,12 @@ class KVCacheRecvingThread(threading.Thread):
             logger.error("Failed to transfer KV cache for request "
                          f"{request_id}: {e}")
         finally:
-            self.task_tracker.update_done_task_count(request_id)
             # Always send the done signal to the remote host to ensure proper
             # resource cleanup. Failing to do so may cause a memory leak on the
             # remote host.
-            self._send_done_recv_signal(request_id, remote_host,
-                                        remote_handshake_port)
+            if offset == num_need_pulls - 1:
+                self._send_done_recv_signal(request_id, remote_host, remote_handshake_port)
+                self.task_tracker.update_done_task_count(request_id)
             self.request_queue.task_done()
 
     def _transfer_kv_cache(self, req_meta: dict[str, Any]):
@@ -313,6 +324,8 @@ class KVCacheRecvingThread(threading.Thread):
         remote_engine_id = req_meta["remote_engine_id"]
         remote_host = req_meta["remote_host"]
         remote_handshake_port = req_meta["remote_handshake_port"]
+        offset = req_meta["offset"]
+        self.num_need_pulls = req_meta["num_need_pulls"]
 
         # Full prefix cache hit: do not need to read remote blocks, just notify
         # P worker that we have the blocks we need.
@@ -321,54 +334,116 @@ class KVCacheRecvingThread(threading.Thread):
 
         # Check if we have the remote metadata cached.
         if remote_engine_id not in self.kv_caches_base_addr or \
-                remote_handshake_port not in self.kv_caches_base_addr[remote_engine_id]:
+            remote_handshake_port not in self.kv_caches_base_addr[remote_engine_id]:
             self._get_remote_metadata(remote_host, remote_handshake_port)
 
-        grouped_remote_block_ids, grouped_local_block_ids = \
-            group_concurrent_contiguous(remote_block_ids, local_block_ids)
+        if self.num_need_pulls == 1:
+            grouped_remote_block_ids, grouped_local_block_ids = \
+                    group_concurrent_contiguous(remote_block_ids, local_block_ids)
+        else:
+            remote_block_ids = list(map(lambda x: [x], remote_block_ids))
+            local_block_ids = list(map(lambda x: [x], local_block_ids))
+            grouped_remote_block_ids, grouped_local_block_ids = remote_block_ids, local_block_ids
+        num_transfer_groups = len(grouped_remote_block_ids)
+
         remote_kv_caches_base_addrs = \
             self.kv_caches_base_addr[remote_engine_id][remote_handshake_port]
         local_kv_caches_base_addrs = \
             self.kv_caches_base_addr[self.local_engine_id][self.local_handshake_port]
+        remote_transfer_port = self.remote_te_port[remote_engine_id][remote_handshake_port]
+        num_blocks = len(local_block_ids)
+        session_id = f"{remote_host}:{remote_transfer_port}"
 
         req_start_time = time.perf_counter()
-        num_transfer_groups = len(grouped_remote_block_ids)
-        num_blocks = len(local_block_ids)
-
-        remote_transfer_port = self.remote_te_port[remote_engine_id][
-            remote_handshake_port]
-        session_id = f"{remote_host}:{remote_transfer_port}"
         src_list, dst_list, length_list = [], [], []
         for k, (src_layer_base_addr, dst_layer_base_addr) in enumerate(
                 zip(local_kv_caches_base_addrs, remote_kv_caches_base_addrs)):
-            block_len = (self.block_len[k % 2]
-                         if self.use_mla else self.block_len[0])
-            for i, remote_block_id in enumerate(grouped_remote_block_ids):
-                local_block_ids = grouped_local_block_ids[i]
-                src = src_layer_base_addr + local_block_ids[0] * block_len
-                dst = dst_layer_base_addr + remote_block_id[0] * block_len
-                length = len(local_block_ids) * block_len
+            block_len = self.block_len[k % 2]
+            inner_block_len = block_len // self.num_need_pulls
+            for remote_block_id, local_block_id in zip(grouped_remote_block_ids, grouped_local_block_ids):
+                src = src_layer_base_addr + local_block_id[0] * block_len + offset * inner_block_len
+                dst = dst_layer_base_addr + remote_block_id[0] * inner_block_len
+                length = inner_block_len * len(local_block_id)
                 src_list.append(src)
                 dst_list.append(dst)
                 length_list.append(length)
-        ret = self.engine.batch_transfer_sync_read(session_id, src_list,
-                                                   dst_list, length_list)
+
+        ret = self.engine.batch_transfer_sync_read(session_id, src_list, dst_list,
+                                                length_list)
         if ret < 0:
-            logger.error("Mooncake transfer failed for request %s",
-                         req_meta["request_id"])
+            logger.error(
+                "Mooncake transfer failed for request %s: ", req_meta["request_id"])
             raise RuntimeError(f"Mooncake transfer failed, ret: {ret}")
 
         req_end_time = time.perf_counter()
         req_transfer_elapsed = (req_end_time - req_start_time) * 1000
-        logger.info(
-            "KV cache transfer for request %s took %.2f ms (%d groups,"
-            " %d blocks).", request_id, req_transfer_elapsed,
-            num_transfer_groups, num_blocks)
+        logger.info("KV cache transfer for request %s took %.2f ms (%d groups,"
+                    " %d blocks). local_ip %s local_device_id %s remote_session_id %s", request_id, req_transfer_elapsed,
+                    num_transfer_groups, num_blocks, get_ip(), self.tp_rank, session_id)
+        if self.num_need_pulls > 1 and offset == self.num_need_pulls -1:
+            self._cat_kv_cache(local_block_ids)
+
+    def _cat_kv_cache(self, block_ids: list[list[int]]):
+        # get necessary params from config
+        head_dim = self.model_config.hf_config.head_dim
+        block_size = self.vllm_config.cache_config.block_size
+        num_kv_head = self.model_config.hf_config.num_key_value_heads
+
+        dtype = list(self.kv_caches.values())[0][0].dtype
+        block_ids = [item for sublist in block_ids for item in sublist]
+        block_ids_tensor = torch.tensor(block_ids, dtype=torch.int32)
+        num_blocks = len(block_ids)
+        block_len = num_blocks * block_size
+        block_table = block_ids_tensor.view(1, -1).npu()
+        block_len_tensor = torch.tensor([block_len], dtype=torch.int32).npu()
+        seq_start_tensor = torch.tensor([0], dtype=torch.int32).npu()
+
+        k_buffer = torch.empty(block_len, num_kv_head, head_dim, dtype=dtype).npu()
+        v_buffer = torch.empty(block_len, num_kv_head, head_dim, dtype=dtype).npu()
+
+        block_offsets = torch.arange(0, block_size, dtype=torch.int32)
+        slot_mapping = block_offsets.reshape(
+            (1, block_size)) + block_ids_tensor.reshape(
+            (num_blocks, 1)) * block_size
+        slot_mapping = slot_mapping.flatten().npu()
+
+        for kv_cache_layer in self.kv_caches.values():
+            k_cache_layer, v_cache_layer = kv_cache_layer
+            torch_npu.atb.npu_paged_cache_load(
+                k_cache_layer,
+                v_cache_layer,
+                block_table,
+                block_len_tensor,
+                seq_starts=seq_start_tensor,
+                key=k_buffer,
+                value=v_buffer,
+            )
+
+            k_buffer = self._transpose_kv_cache_between_head(k_buffer, num_blocks, block_size, block_len, num_kv_head)
+            v_buffer = self._transpose_kv_cache_between_head(v_buffer, num_blocks, block_size, block_len, num_kv_head)
+
+            torch_npu._npu_reshape_and_cache(
+                key=k_buffer,
+                value=v_buffer,
+                key_cache=k_cache_layer,
+                value_cache=v_cache_layer,
+                slot_indices=slot_mapping,
+            )
+        del k_buffer
+        del v_buffer
+
+    def _transpose_kv_cache_between_head(self, buffer: torch.Tensor, num_blocks: int,
+                                         block_size: int, block_len: int, num_kv_head: int) -> torch.Tensor:
+        buffer = buffer.view(num_blocks, self.num_need_pulls, block_size, -1)
+        buffer.transpose_(1, 2)
+        buffer = buffer.contiguous().view(block_len, num_kv_head, -1)
+
+        return buffer
 
     def _get_remote_metadata(self, remote_host: str,
                              remote_handshake_port: int) -> None:
         """Get the metadata from the remote host."""
-        sock: Optional[zmq.Socket] = None  # type: ignore
+        sock: Optional[zmq.Socket] = None
         try:
             sock = self._get_remote_socket(remote_host, remote_handshake_port)
             ensure_zmq_send(sock, self.encoder.encode((GET_META_MSG, "")))
@@ -386,62 +461,51 @@ class KVCacheRecvingThread(threading.Thread):
             if sock is not None:
                 self._return_remote_socket(sock, remote_host,
                                            remote_handshake_port)
-                logger.debug("Returned socket to pool for %s:%d", remote_host,
-                             remote_handshake_port)
+                logger.debug("Returned socket to pool for %s:%d",
+                             remote_host, remote_handshake_port)
 
     def _send_done_recv_signal(self, request_id: str, remote_host: str,
                                remote_handshake_port: int):
         logger.debug("Sending done recving signal for request %s to %s:%d",
                      request_id, remote_host, remote_handshake_port)
-        sock: Optional[zmq.Socket] = None  # type: ignore
+        sock: Optional[zmq.Socket] = None
         try:
             sock = self._get_remote_socket(remote_host, remote_handshake_port)
             data_bytes = self.encoder.encode((DONE_RECVING_MSG, request_id))
             ensure_zmq_send(sock, data_bytes)
-            resp = ensure_zmq_recv(sock,
-                                   self.remote_poller,
+            resp = ensure_zmq_recv(sock, self.remote_poller,
                                    timeout=self.timeout)
-            logger.debug(
-                f"Received response for request {request_id}: {resp.decode('utf-8')}"
-            )
+            logger.debug(f"Received response for request {request_id}: {resp}")
             if resp != b"ACK":
                 logger.error("Failed to receive ACK for request %s from %s:%d",
-                             request_id, remote_host, remote_handshake_port)
-                raise RuntimeError(
-                    f"Failed to receive ACK, resp: {resp.decode('utf-8')}")
+                            request_id, remote_host, remote_handshake_port)
+                raise RuntimeError(f"Failed to receive ACK, resp: {resp}")
         finally:
             if sock is not None:
                 self._return_remote_socket(sock, remote_host,
                                            remote_handshake_port)
-                logger.debug("Returned socket to pool for %s:%d", remote_host,
-                             remote_handshake_port)
+                logger.debug("Returned socket to pool for %s:%d",
+                             remote_host, remote_handshake_port)
 
-    def _get_remote_socket(
-            self, remote_host: str,
-            remote_handshake_port: int) -> zmq.Socket:  # type: ignore
+    def _get_remote_socket(self, remote_host: str,
+                    remote_handshake_port: int) -> zmq.Socket:
         """Get a socket to the remote host."""
         remote_path = make_zmq_path("tcp", remote_host, remote_handshake_port)
         with self.remote_sockets_lock:
             if self.remote_sockets[remote_path]:
                 return self.remote_sockets[remote_path].popleft()
 
-            ctx = zmq.Context()  # type: ignore
-            sock = make_zmq_socket(
-                ctx=ctx,
-                path=remote_path,
-                socket_type=zmq.REQ,  # type: ignore
-                bind=False)
-            sock.setsockopt(
-                zmq.SNDTIMEO,  # type: ignore
-                int(self.timeout * 1000))
-            self.remote_poller.register(sock, zmq.POLLIN)  # type: ignore
+            ctx = zmq.Context()
+            sock = make_zmq_socket(ctx=ctx,
+                                   path=remote_path,
+                                   socket_type=zmq.REQ,
+                                   bind=False)
+            sock.setsockopt(zmq.SNDTIMEO, int(self.timeout * 1000))
+            self.remote_poller.register(sock, zmq.POLLIN)
             return sock
 
-    def _return_remote_socket(
-            self,
-            sock: zmq.Socket,  # type: ignore
-            remote_host: str,
-            remote_handshake_port: int) -> None:
+    def _return_remote_socket(self, sock: zmq.Socket, remote_host: str,
+                              remote_handshake_port: int) -> None:
         """Return the remote socket to the pool."""
         remote_path = make_zmq_path("tcp", remote_host, remote_handshake_port)
         with self.remote_sockets_lock:
@@ -469,7 +533,7 @@ class MooncakeConnectorMetadata(KVConnectorMetadata):
         )
 
 
-class MooncakeConnector(KVConnectorBase_V1):
+class MooncakeConnectorV1_barebone(KVConnectorBase_V1):
 
     def __init__(self, vllm_config: VllmConfig, role: KVConnectorRole):
         assert vllm_config.kv_transfer_config is not None
@@ -561,15 +625,16 @@ class MooncakeConnectorScheduler:
         self.vllm_config = vllm_config
         self.block_size = vllm_config.cache_config.block_size
         self.engine_id = engine_id
+        self.local_ip = get_ip()
         logger.info("Initializing Mooncake Scheduler %s", engine_id)
 
-        self.side_channel_host = get_ip()
+        self.side_channel_host = get_local_ip_by_remote()
         self.max_device_id = vllm_config.parallel_config.tensor_parallel_size * \
                              vllm_config.parallel_config.data_parallel_size
 
         # Handshake base port
         self.side_channel_port = (
-            vllm_config.kv_transfer_config.kv_port +
+            envs_ascend.VLLM_BASE_PORT + 
             vllm_config.parallel_config.data_parallel_rank_local *
             vllm_config.parallel_config.tensor_parallel_size)
 
@@ -605,7 +670,7 @@ class MooncakeConnectorScheduler:
 
         if params is not None and params.get("do_remote_prefill"):
             assert num_computed_tokens == 0, "Currently only support " \
-                                             "prefill with num_computed_tokens == 0."
+                "prefill with num_computed_tokens == 0."
             # Assume that the request's KV cache is already fully prefilled and
             # can be fetched entirely from the prefill node.
             count = max(len(request.prompt_token_ids) - 1, 0)
@@ -701,6 +766,7 @@ class MooncakeConnectorScheduler:
             remote_engine_id=self.engine_id,
             remote_host=self.side_channel_host,
             remote_port=self.side_channel_port,
+            last_token_id=request.output_token_ids[-1],
         )
 
     def get_finished_count(self) -> Optional[int]:
@@ -749,20 +815,20 @@ class MooncakeConnectorWorker:
         self.dp_rank = vllm_config.parallel_config.data_parallel_rank_local
         self.dp_size = vllm_config.parallel_config.data_parallel_size_local
         self.kv_caches: dict[str, torch.Tensor] = {}
-        self.side_channel_host = get_ip()
+        self.side_channel_host = get_local_ip_by_remote()
         self.max_device_id = self.tp_size * self.dp_size
         self.kv_role = vllm_config.kv_transfer_config.kv_role
 
         # Handshake base port
         self.side_channel_port = (
-            vllm_config.kv_transfer_config.kv_port +
+            envs_ascend.VLLM_BASE_PORT +
             vllm_config.parallel_config.data_parallel_rank_local *
             vllm_config.parallel_config.tensor_parallel_size)
         self.handshake_port = self.side_channel_port + self.tp_rank
         self.sockets: dict = {}
 
-        # get tp device id
-        # TODO(kw): https://github.com/vllm-project/vllm-ascend/pull/940
+        # get tp device id 
+        # TODO(kw): https://github.com/vllm-project/vllm-ascend/pull/940 
         # introducing some changes
         device_ids_str = envs_ascend.PHYSICAL_DEVICES
         if device_ids_str is None:
@@ -782,20 +848,26 @@ class MooncakeConnectorWorker:
         assert len(device_ids) > self.tp_rank  # type: ignore
         self.device_id = device_ids[self.tp_rank]  # type: ignore
 
-        if vllm_config.kv_transfer_config.get_from_extra_config(
-                'use_ascend_direct', False):
-            hostname = self.side_channel_host
-        else:
-            hostname = f"{self.side_channel_host}:0:npu_{self.device_id}"
-        self._initialize(hostname=hostname, device_name=None)
+        self._initialize(
+            hostname=self.side_channel_host + ':' + "0" + ':' + 'npu_' \
+                + str(self.device_id),
+            device_name=None)
         self.te_rpc_port = self.engine.get_rpc_port()
 
         # Background thread for sending or receiving KV caches.
         self.kv_send_thread: Optional[KVCacheSendingThread] = None
         self.kv_recv_thread: Optional[KVCacheRecvingThread] = None
 
+        # kv_transfer variables
         self.vllm_config = vllm_config
         self.block_size = vllm_config.cache_config.block_size
+        self.num_key_value_heads = self.vllm_config.model_config.hf_config.num_key_value_heads
+        num_d_block_heads = max(1, self.num_key_value_heads // self.tp_size)
+        num_p_block_heads = max(1, self.num_key_value_heads // self._prefill_tp_size)
+        if self.vllm_config.model_config.is_deepseek_mla:
+            self.num_need_pulls = 1
+        else:
+            self.num_need_pulls = num_d_block_heads // num_p_block_heads
 
     def _get_prefill_decode_size(self, vllm_config: VllmConfig):
         # get prefill tp and dp size from extra config
@@ -825,7 +897,8 @@ class MooncakeConnectorWorker:
     ) -> None:
         """Initialize the mooncake instance."""
         device_name = device_name if device_name is not None else ""
-        ret_value = self.engine.initialize(hostname, "P2PHANDSHAKE", "ascend",
+        ret_value = self.engine.initialize(hostname, "P2PHANDSHAKE",
+                                           envs_ascend.MOONCAKE_CONNECTOR_PROTOCOL,
                                            device_name)
         if ret_value != 0:
             raise RuntimeError(
@@ -838,52 +911,36 @@ class MooncakeConnectorWorker:
         first_kv_cache = first_kv_cache_tuple[0]
 
         # TODO(tms): Find a more robust way to detect and handle MLA
-        self.use_mla = first_kv_cache_tuple[0].size(
-            -1) != first_kv_cache_tuple[1].size(-1)
-        if self.use_mla:
-            # MLA case.[num_block, block_size, 1, hidden_dim]
-            self.num_blocks = first_kv_cache.shape[0]
-            block_rank = 3  # [block_size, latent_dim]
-            block_shape_norm = first_kv_cache_tuple[0].shape[-block_rank:]
-            block_shape_pe = first_kv_cache_tuple[1].shape[-block_rank:]
-            self.block_len = [
-                first_kv_cache[0].element_size() * math.prod(block_shape_norm),
-                first_kv_cache[1].element_size() * math.prod(block_shape_pe)
-            ]
-            logger.info(
-                "num_blocks: %s, block_shape_norm: %s, block_shape_pe: %s",
-                self.num_blocks, block_shape_norm, block_shape_pe)
-        else:
-            # [num_block, block_size, num_head, hidden_dim]
-            self.num_blocks = first_kv_cache.shape[0]
-            kv_elem_size = first_kv_cache.element_size()
-            block_rank = 3  # [block_size, kv_heads, head_dim]
-            block_shape = first_kv_cache.shape[-block_rank:]
-            self.block_len = [kv_elem_size * math.prod(block_shape)]
-            logger.info("num_blocks: %s, block_shape: %s", self.num_blocks,
-                        block_shape)
-
-        logger.info("Registering KV_Caches. use_mla: %s, shape %s",
-                    self.use_mla, first_kv_cache.shape)
+        self.num_blocks = first_kv_cache.shape[0]
+        block_rank = 3
+        kv_elem_size = first_kv_cache.element_size()
+        # MLA case.[num_block, block_size, 1, hidden_dims],
+        # block_shape_1 -> nope, block_shape_2 -> rope
+        # non-MLA case.[num_block, block_size, num_heads, head_dims]
+        # block_shape_1 -> k, block_shape_1 -> v
+        block_shape_1 = first_kv_cache_tuple[0].shape[-block_rank:]
+        block_shape_2 = first_kv_cache_tuple[1].shape[-block_rank:]
+        self.block_len = [
+            kv_elem_size * math.prod(block_shape_1),
+            kv_elem_size * math.prod(block_shape_2)
+        ]
+        logger.info(
+            "num_blocks: %s, block_shape_1: %s, block_shape_2: %s",
+            self.num_blocks, block_shape_1, block_shape_1)
 
         self.kv_caches = kv_caches
         kv_caches_base_addr = []
+        region_len_1 = self.num_blocks * self.block_len[0]
+        region_len_2 = self.num_blocks * self.block_len[1]
         for cache_or_caches in kv_caches.values():
             # Normalize to always be a list of caches
-            if self.use_mla:
-                for i, cache in enumerate(cache_or_caches, 0):
-                    base_addr = cache.data_ptr()
-                    region_len = self.num_blocks * self.block_len[i % 2]
-                    kv_caches_base_addr.append(base_addr)
-                    self._register(base_addr, region_len)
-            else:
-                cache_list = [cache_or_caches
-                              ] if self.use_mla else cache_or_caches
-                for cache in cache_list:
-                    base_addr = cache.data_ptr()
-                    region_len = self.num_blocks * self.block_len[0]
-                    kv_caches_base_addr.append(base_addr)
-                    self._register(base_addr, region_len)
+            assert len(cache_or_caches) == 2, "per-layer kv_caches must equal to 2"
+            base_addr_1 = cache_or_caches[0].data_ptr()
+            base_addr_2 = cache_or_caches[1].data_ptr()
+            kv_caches_base_addr.append(base_addr_1)
+            self._register(base_addr_1, region_len_1)
+            kv_caches_base_addr.append(base_addr_2)
+            self._register(base_addr_2, region_len_2)
 
         # After KV Caches registered, start the sending or receiving thread.
         metadata = MooncakeAgentMetadata(
@@ -895,18 +952,16 @@ class MooncakeConnectorWorker:
 
         ready_event = threading.Event()
         if self.kv_role == 'kv_producer':
-            self.kv_send_thread = KVCacheSendingThread(self.tp_rank,
-                                                       self._decode_tp_size,
-                                                       self.engine_id,
-                                                       self.side_channel_host,
-                                                       self.side_channel_port,
-                                                       metadata, ready_event)
+            self.kv_send_thread = KVCacheSendingThread(
+                self.tp_rank, self._decode_tp_size, self.engine_id,
+                self.side_channel_host, self.side_channel_port, metadata,
+                ready_event, self.kv_caches)
             self.kv_send_thread.start()
         else:
             self.kv_recv_thread = KVCacheRecvingThread(
                 self.tp_rank, self.tp_size, self.engine, self.engine_id,
                 self.handshake_port, kv_caches_base_addr, self.block_len,
-                ready_event)
+                ready_event, self.vllm_config, self.kv_caches)
             self.kv_recv_thread.start()
         ready_event.wait()
 
@@ -919,14 +974,10 @@ class MooncakeConnectorWorker:
             raise RuntimeError("Mooncake memory registration failed.")
 
     def get_finished(self) -> tuple[set[str], set[str]]:
-        done_sending = (
-            self.kv_send_thread.
-            get_and_clear_finished_requests(  # type: ignore[union-attr]
-            ) if self.kv_role == 'kv_producer' else set())
-        done_recving = (
-            self.kv_recv_thread.
-            get_and_clear_finished_requests(  # type: ignore[union-attr]
-            ) if self.kv_role == 'kv_consumer' else set())
+        done_sending = (self.kv_send_thread.get_and_clear_finished_requests()
+                        if self.kv_role == 'kv_producer' else set())
+        done_recving = (self.kv_recv_thread.get_and_clear_finished_requests()
+                        if self.kv_role == 'kv_consumer' else set())
         if self.tp_rank == 0:
             logger.debug(
                 "Number of completed KV cache send requests: %d, receive "
@@ -942,16 +993,19 @@ class MooncakeConnectorWorker:
                 meta.remote_engine_id, len(meta.local_block_ids),
                 len(meta.remote_block_ids))
 
-            remote_handshake_port = meta.remote_port + \
-                                    self._get_remote_tp_rank(req_id)
-            self.kv_recv_thread.add_request(  # type: ignore[union-attr]
-                request_id=req_id,
-                local_block_ids=meta.local_block_ids,
-                remote_block_ids=meta.remote_block_ids,
-                remote_engine_id=meta.remote_engine_id,
-                remote_host=meta.remote_host,
-                remote_handshake_port=remote_handshake_port,
-            )
+            choosen_rank_list = self._get_remote_tp_rank(req_id)
+            remote_handshake_port_list = [x + meta.remote_port for x in choosen_rank_list]
+            for i in range(self.num_need_pulls):
+                self.kv_recv_thread.add_request(
+                    request_id=req_id,
+                    local_block_ids=meta.local_block_ids,
+                    remote_block_ids=meta.remote_block_ids,
+                    remote_engine_id=meta.remote_engine_id,
+                    remote_host=meta.remote_host,
+                    remote_handshake_port=remote_handshake_port_list[i],
+                    offset=i,
+                    num_need_pulls=self.num_need_pulls
+                )
 
         if self.kv_send_thread is not None:
             for req_id, delay_start_time in metadata.requests_to_send.items():
@@ -959,38 +1013,87 @@ class MooncakeConnectorWorker:
                     self.kv_send_thread.add_delayed_request(
                         req_id, delay_start_time)
 
-    def _get_remote_tp_rank(self, req_id: str) -> int:
+    def _get_remote_tp_rank(self, req_id: str) -> List[int]:
         return self._get_remote_tp_ranks_for_req(req_id)[self.tp_rank]
 
-    def _get_remote_tp_ranks_for_req(self, req_id: str) -> list[int]:
+    def _get_remote_tp_ranks_for_req(self, req_id: str) -> List[List[int]]:
         if self._prefill_tp_size == self._decode_tp_size:
-            return list(range(self._prefill_tp_size))
+            result = list(map(lambda x: [x], range(self._prefill_tp_size)))
+            return result
 
         seed = string_to_int64_hash(req_id)
         rand = random.Random(seed)
-        sampled_nums = rand.sample(range(self._prefill_tp_size),
-                                   self._decode_tp_size)
+        sampled_nums = []
+        ori_data = np.arange(self._prefill_tp_size)
+
+        # random split prefill tp list
+        if self._prefill_tp_size > self.num_key_value_heads:
+            # TODO: can be modified
+            num_groups = len(ori_data) // self.num_key_value_heads
+            ori_data = ori_data.reshape(-1, num_groups)
+            rand_group_index = rand.sample(range(num_groups), \
+                max(self.tp_size // self.num_key_value_heads, 1)) # random choose a group
+
+            choosen_group = ori_data[:, [rand_group_index]]
+            flattened = choosen_group.reshape(-1).tolist()
+            sampled_nums = [flattened[i:i + self.num_need_pulls] for i in range(0, len(flattened), self.num_need_pulls)]
+
+        # non-random split
+        else:
+            group_size = self._prefill_tp_size // self.tp_size
+            for i in range(self.tp_size):
+                slice = ori_data[i*group_size : (i+1)*group_size]
+                sampled_nums.append(slice)
         return sampled_nums
 
 
 @contextlib.contextmanager
-def zmq_ctx(socket_type: Any,
-            addr: str) -> Iterator[zmq.Socket]:  # type: ignore
+def zmq_ctx(socket_type: Any, addr: str) -> Iterator[zmq.Socket]:
     """Context manager for a ZMQ socket"""
 
-    if socket_type not in (zmq.ROUTER, zmq.REQ, zmq.DEALER):  # type: ignore
+    if socket_type not in (zmq.ROUTER, zmq.REQ, zmq.DEALER):
         raise ValueError(f"Unexpected socket type: {socket_type}")
 
-    ctx: Optional[zmq.Context] = None  # type: ignore
+    ctx: Optional[zmq.Context] = None
     try:
-        ctx = zmq.Context()  # type: ignore
+        ctx = zmq.Context()
         yield make_zmq_socket(ctx=ctx,
                               path=addr,
                               socket_type=socket_type,
-                              bind=socket_type == zmq.ROUTER)  # type: ignore
+                              bind=socket_type == zmq.ROUTER)
     finally:
         if ctx is not None:
             ctx.destroy(linger=0)
+
+
+def get_local_ip_by_remote() -> str:
+    # try ipv4
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    try:
+        s.connect(("8.8.8.8", 80))  # Doesn't need to be reachable
+        return s.getsockname()[0]
+    except Exception:
+        logger.error("get_local_ip_by_remote Error.")
+
+    try:
+        hostname = socket.gethostname()
+        ip = socket.gethostbyname(hostname)
+        if ip and ip != "127.0.0.1" and ip != "0.0.0.0":
+            return ip
+    except Exception:
+        logger.error("get_local_ip_by_remote Error 2.")
+
+    # try ipv6
+    try:
+        s = socket.socket(socket.AF_INET6, socket.SOCK_DGRAM)
+        # Google's public DNS server, see
+        # https://developers.google.com/speed/public-dns/docs/using#addresses
+        s.connect(
+            ("2001:4860:4860::8888", 80
+             ))  # Doesn't need to be reachable  # Doesn't need to be reachable
+        return s.getsockname()[0]
+    except Exception:
+        raise ValueError("Can not get local ip")
 
 
 def group_concurrent_contiguous(
@@ -1024,33 +1127,26 @@ def string_to_int64_hash(input_str):
     return uint64_value
 
 
-def ensure_zmq_send(
-        socket: zmq.Socket,  # type: ignore
-        data: bytes,
-        max_retries: int = 3):
+def ensure_zmq_send(socket: zmq.Socket, data: bytes, max_retries: int = 3):
     retries_left = max_retries
     while True:
         try:
             socket.send(data)
             return
-        except zmq.ZMQError as e:  # type: ignore
+        except zmq.ZMQError as e:
             retries_left -= 1
             if retries_left > 0:
-                logger.warning(
-                    f"Send failed: {e}, retrying... ({retries_left} "
-                    "attempts left)")
+                logger.warning(f"Send failed: {e}, retrying... ({retries_left} "
+                               "attempts left)")
                 time.sleep(0.1)
             else:
                 logger.error(f"Send failed after all retries: {e}")
                 raise RuntimeError(f"Failed to send data after {max_retries} "
                                    f"retries: {e}")
 
-
-def ensure_zmq_recv(
-        socket: zmq.Socket,  # type: ignore
-        poller: zmq.Poller,  # type: ignore
-        timeout: float = 1.0,
-        max_retries: int = 3) -> bytes:
+           
+def ensure_zmq_recv(socket: zmq.Socket, poller: zmq.Poller,
+                    timeout: float = 1.0, max_retries: int = 3) -> bytes:
     retries_left = max_retries
     while True:
         try:
@@ -1058,8 +1154,8 @@ def ensure_zmq_recv(
                 data = socket.recv()
                 return data
             else:
-                raise zmq.ZMQError("Receive timeout")  # type: ignore
-        except zmq.ZMQError as e:  # type: ignore
+                raise zmq.ZMQError("Receive timeout")
+        except zmq.ZMQError as e:
             retries_left -= 1
             if retries_left > 0:
                 logger.warning(f"Receive failed: {e}, retrying... "
@@ -1067,6 +1163,5 @@ def ensure_zmq_recv(
                 time.sleep(0.1)
             else:
                 logger.error(f"Receive failed after all retries: {e}")
-                raise RuntimeError(
-                    f"Failed to receive data after {max_retries} "
-                    f"retries: {e}")
+                raise RuntimeError(f"Failed to receive data after {max_retries} "
+                                   f"retries: {e}")
